+++ lib/controllers/backpack.js
@@ -9,9 +9,9 @@ const {
   COLOR_ORDER,
   GAMMA_DEFAULT,
   FIRMATA_7BIT_MASK
-} = require('./constants');
-const { Pixel } = require('./index')
-const { create_gamma_table } = require('./utils');
+} = require('../constants');
+const { Pixel } = require('../index')
+const { create_gamma_table } = require('../utils');
 
 const IC2Backpack = {
   initialize: {
