+++ lib/controllers/firmata.js
@@ -12,9 +12,9 @@ const {
   START_SYSEX,
   END_SYSEX,
   PIXEL_COMMAND
-} = require('./constants');
-const { create_gamma_table } = require('./utils');
-const { Pixel } = require('./index')
+} = require('../constants');
+const { create_gamma_table } = require('../utils');
+const { Pixel } = require('../index')
 
 const Firmata = {
   initialize: {
