+++ lib/strip.js
@@ -4,8 +4,8 @@ const { ColorString } = require('./index')
 const {
   SHIFT_FORWARD
 } = require('./constants');
-const { Firmata } = require('./firmata');
-const { IC2Backpack } = require('./backpack');
+const { Firmata } = require('./controllers/firmata');
+const { IC2Backpack } = require('./controllers/backpack');
 
 const strips = new WeakMap();
